moved nymwalletd and nymserverd into zingo-proxyd, behind feature gates
