--- zingo-nymserverd/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "zingo-nymserverd"
-version = "0.1.0"
-edition = "2021"
-
-[[bin]]
-name = "zingonymserverd"
-path = "src/bin/nymserverd.rs"
-
-[lib]
-name = "nymserverlib"
-path = "src/lib.rs"
-
-[features]
-nym_wallet = []
-
-[dependencies]
-zingo-rpc = { path = "../zingo-rpc" }
-
-http = "0.2.4"
-tokio = { version = "1", features = ["full"] }
-tonic = "0.10.2"
-# zcash_client_backend = { git = "https://github.com/zingolabs/librustzcash.git", tag = "zingo_rc.2.1", features = ["lightwalletd-tonic"] }
-zcash_client_backend = { git = "https://github.com/zingolabs/librustzcash.git", branch = "nym_integration", features = ["lightwalletd-tonic"] }
-# zingo-netutils = { git = "https://github.com/aloearev/zingolib.git", branch = "cratify_netutils" }
-zingo-netutils = { git = "https://github.com/zingolabs/zingolib.git", branch = "nym_integration" }
-prost = "0.12"
-bytes = "1.1"
-tokio-socks = "0.5"
-nym-sdk = { git = "https://github.com/nymtech/nym", branch = "master" }
-nym-sphinx-addressing = { git = "https://github.com/nymtech/nym", branch = "master" }
-nym-validator-client = { git = "https://github.com/nymtech/nym", branch = "master" }
-nym-bin-common = { git = "https://github.com/nymtech/nym", branch = "master" }
-nym-sphinx-anonymous-replies = { git = "https://github.com/nymtech/nym", branch = "master" }
-ctrlc = "3.2.1"
-
