--- zingo-nymwalletd/src/lib.rs
@@ -1,6 +1,0 @@
-//! Nym-Proxy gRPC server
-
-#![warn(missing_docs)]
-#![forbid(unsafe_code)]
-
-pub mod server;
