--- zingo-nymwalletd/zebrad.toml
@@ -1,2 +1,0 @@
-network.network = "Testnet"
-rpc.listen_addr = '127.0.0.1:18232'
