+++ zingo-proxyd/Cargo.toml
@@ -11,6 +11,10 @@ path = "src/bin/zingoproxyd.rs"
 name = "zingoproxylib"
 path = "src/lib.rs"
 
+[features]
+nym_wallet = []
+nym_server = []
+
 [dependencies]
 zingo-rpc = { path = "../zingo-rpc" }
 
@@ -21,3 +25,12 @@ tonic = "0.10.2"
 zcash_client_backend = { git = "https://github.com/zingolabs/librustzcash.git", branch = "nym_integration", features = ["lightwalletd-tonic"] }
 # zingo-netutils = { git = "https://github.com/aloearev/zingolib.git", branch = "cratify_netutils" }
 zingo-netutils = { git = "https://github.com/zingolabs/zingolib.git", branch = "nym_integration" }
+prost = "0.12"
+bytes = "1.1"
+tokio-socks = "0.5"
+nym-sdk = { git = "https://github.com/nymtech/nym", branch = "master" }
+nym-sphinx-addressing = { git = "https://github.com/nymtech/nym", branch = "master" }
+nym-validator-client = { git = "https://github.com/nymtech/nym", branch = "master" }
+nym-bin-common = { git = "https://github.com/nymtech/nym", branch = "master" }
+nym-sphinx-anonymous-replies = { git = "https://github.com/nymtech/nym", branch = "master" }
+ctrlc = "3.2.1"
