+++ zingo-proxyd/src/lib.rs
@@ -1,6 +1,8 @@
-//! Zingo-Proxy gRPC Server
+//! Zingo-Proxy gRPC Server.
+//! NOTE: This is currently a very simple implementation meant only for development and testing, and in its current form should not be used to run mainnet nodes.
 
 #![warn(missing_docs)]
 #![forbid(unsafe_code)]
 
+pub mod nym_server;
 pub mod server;
